+++ emspayafterpay/emspayafterpay.php
@@ -338,7 +338,7 @@ class emspayAfterpay extends EmsPayPaymentModule
     {
         $customer = $this->createCustomer($cart, $locale);
         try {
-            $response = $this->ginger->createOrder([
+            $response = $this->ginger->createOrder(array_filter([
 		    'amount' => Helper::getAmountInCents($cart->getOrderTotal(true)),   // Amount in cents
 		    'currency' => $this->getPaymentCurrency(),                          // Currency
 		    'transactions' => [
@@ -353,7 +353,7 @@ class emspayAfterpay extends EmsPayPaymentModule
 		    'extra' => ['plugin' => $this->getPluginVersion()],                 // Extra information
 		    'webhook_url' => $this->getWebhookUrl(),                            // Webhook URL
 		    'order_lines' => $this->getOrderLines($cart)                        // Order lines
-            ]);
+            ]));
         } catch (\Exception $exception) {
             return Tools::displayError($exception->getMessage());
         }