+++ emspayideal/emspayideal.php
@@ -96,7 +96,7 @@ class emspayIdeal extends EmsPayPaymentModule
     {
 	  $customer = $this->createCustomer($cart, $locale);
 	  try {
-		$response = $this->ginger->createOrder([
+		$response = $this->ginger->createOrder(array_filter([
 		    'amount' => Helper::getAmountInCents($cart->getOrderTotal(true)),   // Amount in cents
 		    'currency' => $this->getPaymentCurrency(),                          // Currency
 		    'transactions' => [
@@ -112,7 +112,7 @@ class emspayIdeal extends EmsPayPaymentModule
 		    'customer' => $customer->toArray(),                                 // Customer information
 		    'extra' => ['plugin' => $this->getPluginVersion()],                 // Extra information
 		    'webhook_url' => $this->getWebhookUrl(),                            // Webhook URL
-		]);
+		]));
         } catch (\Exception $exception) {
             return Tools::displayError($exception->getMessage());
         }